# Vectorize large palette lookup