--- a/config.m4
+++ b/config.m4
@@ -8,7 +8,7 @@ if test "$PHP_CHUNKUTILS2" != "no"; then
   PHP_REQUIRE_CXX()
 
   dnl the 6th parameter here is required for C++ shared extensions
-  PHP_NEW_EXTENSION(chunkutils2, chunkutils2.cpp src/PhpLightArray.cpp src/PhpPalettedBlockArray.cpp src/PhpSubChunkConverter.cpp, $ext_shared,, -DZEND_ENABLE_STATIC_TSRMLS_CACHE=1 -std=c++14 -fno-strict-aliasing -DGSL_THROW_ON_CONTRACT_VIOLATION=1, yes)
+  PHP_NEW_EXTENSION(chunkutils2, chunkutils2.cpp src/PhpLightArray.cpp src/PhpPalettedBlockArray.cpp src/PhpSubChunkConverter.cpp, $ext_shared,, -DZEND_ENABLE_STATIC_TSRMLS_CACHE=1 -std=c++14 -fno-strict-aliasing -ftree-vectorize -DGSL_THROW_ON_CONTRACT_VIOLATION=1, yes)
   PHP_SUBST(CHUNKUTILS2_SHARED_LIBADD)
   PHP_ADD_BUILD_DIR($ext_builddir/src, 1)
   PHP_ADD_BUILD_DIR($ext_builddir/lib, 1)