@dktapps 2024-12-17
While this optimization is nice, introducing maps for lookups on 8 and 16 bpb produced much larger performance gains:
```
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
read<VanillaPaletteSize::BPB_0>          826 ns          826 ns       779054
read<VanillaPaletteSize::BPB_1>         2664 ns         2664 ns       263900
read<VanillaPaletteSize::BPB_2>         2677 ns         2677 ns       263399
read<VanillaPaletteSize::BPB_3>         3661 ns         3661 ns       191157
read<VanillaPaletteSize::BPB_4>         3133 ns         3133 ns       229104
read<VanillaPaletteSize::BPB_5>         3675 ns         3675 ns       191470
read<VanillaPaletteSize::BPB_6>         3701 ns         3701 ns       187440
read<VanillaPaletteSize::BPB_8>         2796 ns         2796 ns       249380
read<VanillaPaletteSize::BPB_16>        3501 ns         3501 ns       199715
write<VanillaPaletteSize::BPB_0>       0.000 ns        0.000 ns   1000000000000
write<VanillaPaletteSize::BPB_1>        6724 ns         6724 ns       102676
write<VanillaPaletteSize::BPB_2>        6338 ns         6338 ns       111334
write<VanillaPaletteSize::BPB_3>        9610 ns         9610 ns        72295
write<VanillaPaletteSize::BPB_4>       12318 ns        12318 ns        58291
write<VanillaPaletteSize::BPB_5>       19162 ns        19162 ns        35687
write<VanillaPaletteSize::BPB_6>       38124 ns        38124 ns        18199
write<VanillaPaletteSize::BPB_8>        9226 ns         9226 ns        76319
write<VanillaPaletteSize::BPB_16>       9937 ns         9937 ns        69880
```


This benchmark used maps for lookups on 8 and 16 bpb, which produced a significantly larger performance improvement than using vectorization (almost bringing the performance in line with smaller palettes).